Fix warning

+++ src/ArduinoJson/MsgPack/endianess.hpp
@@ -17,14 +17,14 @@ inline void swap(T& a, T& b) {
   b = t;
 }
 
-inline void fixEndianess(uint8_t* p, integral_constant<int, 8>) {
+inline void fixEndianess(uint8_t* p, integral_constant<size_t, 8>) {
   swap(p[0], p[7]);
   swap(p[1], p[6]);
   swap(p[2], p[5]);
   swap(p[3], p[4]);
 }
 
-inline void fixEndianess(uint8_t* p, integral_constant<int, 4>) {
+inline void fixEndianess(uint8_t* p, integral_constant<size_t, 4>) {
   swap(p[0], p[3]);
   swap(p[1], p[2]);
 }
@@ -33,7 +33,7 @@ template <typename T>
 inline void fixEndianess(T& value) {
 #if ARDUINOJSON_LITTLE_ENDIAN
   fixEndianess(reinterpret_cast<uint8_t*>(&value),
-               integral_constant<int, sizeof(T)>());
+               integral_constant<size_t, sizeof(T)>());
 #endif
 }
